Change BsmProperties msgCnt and secMark to Integer

+++ jpo-geojsonconverter/src/main/java/us/dot/its/jpo/geojsonconverter/pojos/geojson/bsm/BsmProperties.java
@@ -85,8 +85,8 @@ public class BsmProperties {
     private BrakeSystemStatus brakes;
     private Double heading;
     private String id;
-    private Long msgCnt;
-    private Long secMark;
+    private Integer msgCnt;
+    private Integer secMark;
     private VehicleSize size;
     private Double speed;
     private TransmissionState transmission;
